chrome-ai-extension-builder/assets/templates/backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI/LLM
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.8.0

# HTTP client for tools
requests>=2.31.0
httpx>=0.26.0

# Production (optional)
# redis>=5.0.0
# gunicorn>=21.0.0
